flask-api/requirements.txt
Flask == 1.1.2
fastapi == 0.63.0
Flask-PyMongo==2.3.0
pydantic==1.9.0
mongomock==3.23.0